requirements.txt
CairoSVG==2.4.2
Pillow==8.3.1
python_igraph==0.9.6
